Moved generatePDF and parseSpamAnalysis to spamassassin.ts
