+++ src/lib/save.ts
@@ -1,6 +1,5 @@
-import { readFileSync, writeFileSync, createWriteStream } from 'node:fs';
+import { readFileSync, writeFileSync } from 'node:fs';
 import { __dirname} from '../api/filesystem.js';
-import PDFDocument from 'pdfkit';
 import Cryptr from 'cryptr';
 
 export interface AppState {
@@ -112,94 +111,4 @@ export function save(type: 'paths' | 'config', key: string, value: string): void
 
   const infoString: string = JSON.stringify(info, null, 2);
   writeFileSync(__dirname + `config\\${type}.json`, infoString);
-}
-
-interface SpamAnalysis {
-  [rule: string]: string;
-}
-
-interface SpamResult {
-  totalPoints: number;
-  analysis: SpamAnalysis;
-}
-
-/**
- * @description Parses the spam analysis from the log.txt file that SpamAssassin generates
- *
- * @remarks
- * This function takes the output that SpamAssassin generates and parses it into a
- * SpamResult object. The SpamResult object contains the total points and a
- * SpamAnalysis object that contains the rule and description for each rule.
- * The SpamAnalysis object is a dictionary where the key is the rule and the
- * value is the description.
- *
- * @example
- *
- * // Returns { totalPoints: 5.1, analysis: { 'BAYES_50': 'BODY: Bayes spam probability is 50 to 60%'... } }
- * const spamResult = parseSpamAnalysis(emailText);
- *
- * @param {string} emailText - The log.txt file that SpamAssassin generates
- * @returns {SpamResult | null} - The result of the spam analysis
- */
-export function parseSpamAnalysis(emailText: string): SpamResult | null {
-  const startIndex: number = emailText.indexOf('Content analysis details:');
-
-  const analysisText: string = emailText.substring(startIndex);
-  const analysisLines: string[] = analysisText.split('\n').map(line => line.trim());
-
-  const analysis: SpamAnalysis = {};
-  let totalPoints: number = 0;
-
-  for (const line of analysisLines) {
-    const match: RegExpMatchArray | null = line.match(/^([\d.-]+)\s+(\w+)\s+(.*)/);
-    if (match) {
-      const [, pointsString, rule, description] = match;
-      const points = parseFloat(pointsString);
-      analysis[rule] = description;
-      totalPoints += points;
-    }
-  }
-
-  totalPoints = Number(totalPoints.toFixed(1));
-
-  return {
-    totalPoints,
-    analysis,
-  };
-}
-
-/**
- * @description Generates a PDF file from a SpamResult object
- *
- * @remarks
- * This function takes a SpamResult object and generates a PDF file using PDFKit.
- * It then saves the PDF file to the user's desktop.
- *
- * @example
- *
- * // Generates and saves a PDF file
- * generatePDF(spamResult);
- *
- * @param {SpamResult} spamResult - The result of the spam analysis
- */
-export function generatePDF(spamResult: SpamResult): void {
-  const doc = new PDFDocument();
-  const filePath = __dirname + 'temp\\report.pdf';
-
-  doc.fontSize(25).text('Spam Analysis', {
-    underline: true,
-  });
-
-  doc.fontSize(15).text(`Total Points: ${spamResult.totalPoints}`);
-
-  doc.fontSize(15).text('Analysis:', {
-    underline: true,
-  });
-
-  Object.keys(spamResult.analysis).forEach((key) => {
-    doc.fontSize(15).text(`${key}: ${spamResult.analysis[key]}`);
-  });
-
-  doc.pipe(createWriteStream(filePath));
-  doc.end();
 }
